processing/guf_urban_footprint/requirements.txt
rasterio>=1.3.0
requests>=2.28.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
rio-cogeo>=5.0.0
